# walking-safely/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Walking Safely API",
        "description": "API para o aplicativo Walking Safely - Navegação segura com análise de risco em tempo real",
        "contact": {
            "name": "Walking Safely Team",
            "email": "suporte@walkingsafely.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor de Desenvolvimento"
        },
        {
            "url": "https://api.walkingsafely.com/api",
            "description": "Servidor de Produção"
        }
    ],
    "paths": {
...
        "/heatmap": {
            "get": {
                "tags": [
                    "Heatmap"
                ],
                "summary": "Obter dados do mapa de calor",
                "description": "Retorna dados de concentração de ocorrências para renderização de mapa de calor. Granularidade ajusta automaticamente com o zoom.",
                "operationId": "getHeatmap",
                "parameters": [
                    {
                        "name": "min_lat",
                        "in": "query",
                        "description": "Latitude mínima do bounding box",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": -23.6
                        }
                    },
                    {
                        "name": "min_lng",
                        "in": "query",
                        "description": "Longitude mínima do bounding box",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": -46.7
                        }
                    },
                    {
                        "name": "max_lat",
                        "in": "query",
                        "description": "Latitude máxima do bounding box",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": -23.5
                        }
                    },
                    {
                        "name": "max_lng",
                        "in": "query",
                        "description": "Longitude máxima do bounding box",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": -46.6
                        }
                    },
                    {
                        "name": "zoom",
                        "in": "query",
                        "description": "Nível de zoom do mapa (1-18)",
                        "schema": {
                            "type": "integer",
                            "maximum": 18,
                            "minimum": 1,
                            "example": 12
                        }
                    },
                    {
                        "name": "crime_type_id",
                        "in": "query",
                        "description": "Filtrar por tipo de crime",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "crime_category_id",
                        "in": "query",
                        "description": "Filtrar por categoria de crime",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Data inicial",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Data final",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "days",
                        "in": "query",
                        "description": "Últimos N dias",
                        "schema": {
                            "type": "integer",
                            "maximum": 365,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do mapa de calor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/HeatmapPoint"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/heatmap/regions": {
            "get": {
                "tags": [
                    "Heatmap"
                ],
                "summary": "Mapa de calor por região",
                "description": "Retorna dados de ocorrências agregados por região administrativa",
                "operationId": "getHeatmapByRegion",
                "parameters": [
                    {
                        "name": "crime_type_id",
                        "in": "query",
                        "description": "Filtrar por tipo de crime",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "crime_category_id",
                        "in": "query",
                        "description": "Filtrar por categoria de crime",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Data inicial",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Data final",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "days",
                        "in": "query",
                        "description": "Últimos N dias",
                        "schema": {
                            "type": "integer",
                            "maximum": 365,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados agregados por região",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/RegionHeatmap"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/heatmap/distribution": {
            "get": {
                "tags": [
                    "Heatmap"
                ],
                "summary": "Distribuição por tipo de crime",
                "description": "Retorna a distribuição de ocorrências por tipo de crime",
                "operationId": "getDistribution",
                "parameters": [
                    {
                        "name": "min_lat",
                        "in": "query",
                        "description": "Latitude mínima do bounding box",
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "min_lng",
                        "in": "query",
                        "description": "Longitude mínima do bounding box",
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "max_lat",
                        "in": "query",
                        "description": "Latitude máxima do bounding box",
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "max_lng",
                        "in": "query",
                        "description": "Longitude máxima do bounding box",
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Data inicial",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Data final",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "days",
                        "in": "query",
                        "description": "Últimos N dias",
                        "schema": {
                            "type": "integer",
                            "maximum": 365,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Distribuição por tipo de crime",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CrimeDistribution"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HeatmapPoint": {
                "properties": {
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": -23.5505
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": -46.6333
                    },
                    "weight": {
                        "type": "number",
                        "format": "float",
                        "example": 0.75
                    },
                    "count": {
                        "type": "integer",
                        "example": 15
                    }
                },
                "type": "object"
            },
            "RegionHeatmap": {
                "properties": {
                    "region_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "region_name": {
                        "type": "string",
                        "example": "Centro"
                    },
                    "occurrence_count": {
                        "type": "integer",
                        "example": 150
                    },
                    "risk_index": {
                        "type": "number",
                        "format": "float",
                        "example": 0.72
                    },
                    "centroid": {
                        "properties": {
                            "latitude": {
                                "type": "number",
                                "format": "float"
                            },
                            "longitude": {
                                "type": "number",
                                "format": "float"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "CrimeDistribution": {
                "properties": {
                    "crime_type_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "crime_type_name": {
                        "type": "string",
                        "example": "Assalto"
                    },
                    "count": {
                        "type": "integer",
                        "example": 250
                    },
                    "percentage": {
                        "type": "number",
                        "format": "float",
                        "example": 35.5
                    }
                },
                "type": "object"
            },
...
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Os dados fornecidos são inválidos"
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter your Bearer token in the format: Bearer {token}"
            }
        }
    },
...
}